requirements.txt
# ===========================================================================
# PDF Converter Bot - Python Dependencies
# ===========================================================================
# Install: pip install -r requirements.txt
# ===========================================================================

# ---------------------------------------------------------------------------
# Web Framework & Bot
# ---------------------------------------------------------------------------
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-telegram-bot>=20.7
pydantic-settings>=2.0.0
python-multipart>=0.0.9

# ---------------------------------------------------------------------------
# Database (PostgreSQL + Async SQLAlchemy)
# ---------------------------------------------------------------------------
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# ---------------------------------------------------------------------------
# Task Queue (Celery + Redis)
# ---------------------------------------------------------------------------
celery[redis]>=5.3.0
redis>=5.0.0

# ---------------------------------------------------------------------------
# File Processing
# ---------------------------------------------------------------------------
# PyMuPDF для работы с PDF
PyMuPDF>=1.23.0
# pdfplumber для извлечения таблиц
pdfplumber>=0.10.0
# Обработка изображений
Pillow>=10.0.0
# OCR (требуется системный tesseract)
pytesseract>=0.3.10

# ---------------------------------------------------------------------------
# Office Formats (через Gotenberg/LibreOffice)
# ---------------------------------------------------------------------------
# python-docx для постобработки DOCX
python-docx>=1.1.0
# openpyxl для постобработки XLSX
openpyxl>=3.1.0

# ---------------------------------------------------------------------------
# Security & Validation
# ---------------------------------------------------------------------------
# Валидация MIME-типов
python-magic>=0.4.27
# Rate limiting
slowapi>=0.1.9
# JWT аутентификация
python-jose[cryptography]>=3.3.0
# Хэширование паролей
passlib[bcrypt]>=1.7.4

# ---------------------------------------------------------------------------
# Storage (MinIO / S3)
# ---------------------------------------------------------------------------
boto3>=1.34.0

# ---------------------------------------------------------------------------
# HTTP Client (for Gotenberg API)
# ---------------------------------------------------------------------------
httpx>=0.26.0

# ---------------------------------------------------------------------------
# Monitoring
# ---------------------------------------------------------------------------
prometheus-client>=0.19.0

# ---------------------------------------------------------------------------
# Testing (development only)
# ---------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
